moved dymension to separate file
